requirements.txt
asgiref==3.8.1
build==1.2.2.post1
click==8.1.8
Django==5.1.7
packaging==24.2
pip-tools==7.4.1
pyproject_hooks==1.2.0
ruff==0.11.2
setuptools==77.0.3
sqlparse==0.5.3
wheel==0.45.1
appdirs==1.4.4
asgiref==3.8.1
boto3==1.37.18
botocore==1.37.18
build==1.2.2.post1
click==8.1.8
Django==5.1.7
django-bakery==0.13.5
fs==2.4.16
jmespath==1.0.1
packaging==24.2
pip-tools==7.4.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
ruff==0.11.2
s3transfer==0.11.4
setuptools==77.0.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
wheel==0.45.1
appdirs==1.4.4
asgiref==3.8.1
boto3==1.37.18
botocore==1.37.18
build==1.2.2.post1
click==8.1.8
colorama==0.4.6
cssbeautifier==1.15.4
Django==5.1.7
djlint==1.36.4
EditorConfig==0.17.0
fs==2.4.16
jmespath==1.0.1
jsbeautifier==1.15.4
json5==0.10.0
packaging==24.2
pathspec==0.12.1
pip-tools==7.4.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
ruff==0.11.2
s3transfer==0.11.4
setuptools==77.0.3
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
urllib3==2.3.0
wheel==0.45.1
appdirs==1.4.4
asgiref==3.8.1
boto3==1.37.18
botocore==1.37.18
build==1.2.2.post1
click==8.1.8
colorama==0.4.6
cssbeautifier==1.15.4
Django==5.1.7
django-bakery==0.13.5
djlint==1.36.4
dotenv==0.9.9
EditorConfig==0.17.0
fs==2.4.16
jmespath==1.0.1
jsbeautifier==1.15.4
json5==0.10.0
packaging==24.2
pathspec==0.12.1
pip-tools==7.4.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
ruff==0.11.2
s3transfer==0.11.4
setuptools==77.0.3
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
urllib3==2.3.0
wheel==0.45.1
appdirs==1.4.4
asgiref==3.8.1
boto3==1.37.18
botocore==1.37.18
build==1.2.2.post1
click==8.1.8
colorama==0.4.6
cssbeautifier==1.15.4
Django==5.1.7
django-bakery==0.13.5
djlint==1.36.4
dotenv==0.9.9
EditorConfig==0.17.0
fs==2.4.16
jmespath==1.0.1
jsbeautifier==1.15.4
json5==0.10.0
packaging==24.2
pathspec==0.12.1
pip-tools==7.4.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
ruff==0.11.2
s3transfer==0.11.4
setuptools==77.0.3
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
urllib3==2.3.0
wheel==0.45.1
